Add encode_quality_string, default stop/start
This adds two functions for encoding quality score vectors as
PHRED-encoded strings (in-place and into a new vector). It also modifies
the `start` and `stop` parameters to default to the entire length of the
input sequence, or in the case of in-place functions, to the minimum of
the lengths of the input and output vectors.