Clarify status of project in readme
I think some people are conflating Bio.jl with the whole BioJulia project, and might be put off by the warning if they land on the Bio.jl repo. See for instance https://discourse.julialang.org/t/finding-new-packages-and-docs/10417/7.

This clarifies the status of the project. I will merge in a day unless anyone objects.